save workspace (tabs/buffers) in session
tabspace does not seem to respect which buffers are opened,
nor their configuration within the window.

desktop.el does respect buffer configurations but pollutes
the scoped buffers of tabspaces.

remove tabspace sessions and rely on desktop to provide tab/buffer
session.

NOTE: when starting tabspaces we reset the scoped buffer but it
only works for the current active tab